Simplify parameter fetching
Number of parameters can be get from the length of the fetched
parameter sequence, so we don't actually need fetch_NPAR methods.

We don't need casting of the returned value into a list, any
sequence returned by the method should be fine. BTW, we still need
exception handiling, to be implemented in a future commits